Another Major update

+++ Pricer/OrderBook/Order.h
@@ -8,7 +8,7 @@ struct Order
 	char m_orderType;//Add or Reduce
 	std::string m_orderId;//unique identifier for an order
 	char m_orderAction;//previously m_side. Buy or Sell
-	double m_price;
+	long long m_price;
 	int m_size;
 	
 	Order();
